+++ packages/core/src/AppCard/AppName.tsx
@@ -2,7 +2,11 @@ import { Heading } from "../index";
 import useAppCard from "./AppCardContext";
 
 const AppName = () => {
-  const { isHeading, level, headingStyle, displayName } = useAppCard();
+  const { isHeading, level, headingStyle, displayName, variant } = useAppCard();
+
+  if (variant === "square") {
+    <p>{displayName}</p>;
+  }
 
   return isHeading ? (
     <Heading {...{ level, headingStyle }}>{displayName}</Heading>